Fix issue 80: new secret not auto selected in dropdown.
Problem was caused because on creation of the new secret, the modal was
toggled to no longer be shown and as part of that function, the
'newSecretName' was reset to ''.  The subsequent action to then set
'gitsecret' in state, simply set the value to '' and thus it was not
shown as the 'selectedItem'.

This change removes the reset of the 'newSecretName' during toggling of
the modal.  We do not need to worry about the name being shown again if
the user subsequently clicked on the plus icon, as this 'newSecretName'
is set to '' after 'gitsecret' has been set in state, after the modal
is toggled.