Multiple pipelines on a repository
Adding ability to have multiple pipelines trigger off of one single
webhook on the actual git repository.  Currently you "add" multiple
webhooks in the tekton UI but only one is actually created in
github.

The monitor task has also been modified to modify all the resultant
pipelineruns that are created.

The comment added to pull requests has been updated into a tabular
format with links into the dashboard.

The pull request status is now set to "pending" once the monitor
task begins running, it is updated to one of "success", "failure"
or "error" depending on the result of the pipelineruns.

All pipelineruns must succeed for "success" to be registered.

New documentation has been delivered and existing documentation
modified to be more pleasing to the eye.