Use correct PipelineRun URL in status table for OpenShift
The status table added to the pull request after pipelineruns have completed,
includes a link to the pipelinerun from the table. On openshift this link
does not appear to be obtaining the correct route - this change uses the
internal dashboard service rather than the https dashboard service to exploit
using the endpoint REST endpoint to get the correct external route to the
dashboard.