Change data used from webhook payload in pipelineresources
The pull request pipeline resource appears to want the HTMLURL from the
GitHub webhooks payload, otherwise the string manipulation fails to find
the pull request number.

Additionally, the way the branch name is obtained in the sink needs to
differ between push and pull-request events as the payloads are different.