Add Octant plugin
This adds a simple plugin for
[Octant](https://github.com/vmware/octant), a new extensible Kubernetes
dashboard.

Out of the box, Octant displays high-level information about built-in
Kubernetes resources like Pods, and minimal information it can derive
about custom resource types like ours.

This plugin adds deeper information about Tekton resources, such as
a link to a TaskRun's underlying Pod, which also surfaces that Pod's
logs.

Octant provides built-in support for Graphviz visualizations, which
could be useful for visualizing the dependency graph of a Pipeline, and
the execution status of a PipelineRun.