TLS certificate is added to the ingress for the eventlistener
when the callback URL has "https" protocol.
The name of the TLS certificate secret can be specified in WEBHOOK_TLS_CERTIFICATE
environment variable.  The default is "cert-" + eventlistener name.
When the secret doesn't exit with the name, the new certificate secret
is created and the certificate is signed by the framework.