Add a yaml parser, a GenerateTask function and unit tests
This commit is the first step of Tekton Generators. It creates a yaml
parser to read the file form Stdin and store the result in the
self-defined struct and then create a unit test to test it.

It also adds the feature to generate the given simple Github spec to
the Tekton Task. This initial version only generates steps and
workspaces, leaving out other fields for further implementation.