Add support to generate pipelinerun
Add the function to generate pipelinerun in the generators. Change to
use github.Spec.Revision in the pipeline params, because the param
substitution does not work properly in the finally. It can be removed
when this bug is fixed.

Update the comment describing the furture change of parameters in the
finally. For each generated resource function, add the validation check
of the input configuration.

In this commit I change the param in the pipeline as hard-coded value,
because param substitution isn't working properly in the finally. I add
the link of that issue to reference so that when it is resolved,
StringVal of SHA can be changed back to $(params.gitrevision)