Add new PipelineRun subcommand
This commit is to add new PipelineRun subcommand to Generators cli. This
is to provide users with a new way to run their generated pipeline.
Change the pkg/write.go because adding the new function to write pipelinerun
and trigger separately and also extracting the similar codes to
functions like parseFile(), getResource()