Hub: return actual modified time of resource
Previously, API sync service used to use update_at time to infer the
date a resource was last modified. The sync service also used to parse
the entire catalog if requested even if it was already parsed and
updated in DB.

This patch fixes it by exiting early if the SHA/HEAD of the catalog last
scanned is the same. It also adds a modified_at field to compute and
store the last modified time of the resource instead of relying on the
updated_at so that even if the catalog gets updated, the modified_at
still remains the same.

Signed-off-by: Sunil Thaha <sthaha@redhat.com>