Results API - use new TaskRun neutral Results API.
This results API is more execution type agnostic, which will help us support
PipelineRuns and other types (e.g. DSL) in the future.

This is largely a refactoring change, with some minor changes and
cleanups made along the way:
- Use simpler test structs for UpdateResult test cases.
- Adds post-update merge validation to make sure identifiers are not
  mutated.
- Modifies tests to use API calls like Get/List instead of direct DB queries for reading result data.

For now, only TaskRuns are still supported for querying.
Things that will come in future commits:
- Parent-scoped ListResults.
- PipelineRun querying.