Results: Configure API Server and Watcher to run against remote servers.
Largely a rehash of the work Jason did in #539, but focuses on any
remote DB, not just one running locally in the cluster.

- Configures API Server to run against an remote MySQL database.
  - Primarily uses environment variables instead of flags for
    configuration, since this just happened to be easier. I'm not opposed
    to moving this back to flags at a later point in time.
- Configures Result Watcher to dial into remote API Server using TLS +
  authentication credentials.
- Slightly modifies storage schema to make it MySQL friendly (still
  uses sqlite for local unit tests)

With these, I was able to get a working e2e demo running, with the
API Server running on Cloud Run + Cloud SQL and the watcher running
locally.