[notifiers] GitHub App Notifier
This controller listens for Tekton TaskRuns that are annotated with
specific integration values, and converts the TaskRun into a GitHub
CheckRun with corresponding status and logs.

Still a WIP, things still to do:
- Add k8s deployment configs
- Add settings to control behavior:
  - What gets published (e.g. should logs always be published?)
  - Installation restrictions (e.g. currently you can trick the controller to
    publish to any installation of the app by manually mutating the annotation,
    ideally we would have an ACL somewhere that limits the orgs/users a
    given namespace can access).

This does not do anything to grant the running TaskRun access to the App
credentials (e.g. for cloning the repo) -
https://github.com/tektoncd/catalog/tree/master/task/github-app-token/0.1
can be used for this purpose for the time being (though note this
doesn't support fine-grained installation permissions).