[results] Implements the ETAG to solve the race condition problem
There might be multiple update requests for one result in the same
period, which could lead to the race condition problem.

To avoid that, we need to make sure the result in our update request
untouched before being updated to the database.  We achieve this by
the `etag`. A valid update request must carry the ETag of the result
that to be updated.

We encode the serialized result to a string as its etag. The etag is
like a summary of a result's content. Once the content changed, the
etag would also be different. For more information, please check here:
https://developers.google.com/gdata/docs/2.0/reference#ResourceVersioning