Common Expression Language Custom Task
This is the initial of an experimental project that provides support for
Common Expression Language (CEL) in Tekton Pipelines.

The functionality is provided by a controller that implements the
Custom Task interface. Its use cases include evaluating complex
expressions to be used in [`WhenExpressions`](https://github.com/tektoncd/pipeline/blob/master/docs/pipelines.md#guard-task-execution-using-whenexpressions)
in subsequent `Tasks` to guard their execution.

The `CEL` expressions to be evaluated by the `CELRun` are specified
using parameters.

As the `Run` executes, its `status` field accumulates information about
the execution status of the `Run` in general. If the evaluation is
successful, it will also contain the results of the evaluation under
`status.results` with the corresponding names of the CEL expressions as
provided in the parameters.