Pipelines In Pipelines
This is an experimental project to provide support for `Pipelines` in
`Pipelines` to improve the composability and reusability of
[Tekton Pipelines](https://github.com/tektoncd/pipeline).

The functionality is provided by a controller
that implements the `Custom Task` interface.

Its use cases include enabling defining and executing a set of `Tasks`
as a complete unit of execution and decoupling `Pipeline` branches
failures such that a `Task` failure in one branch does not stop
execution of unrelated `Pipeline` branches.

Experimental project is proposed in https://github.com/tektoncd/community/issues/330
Issue is further described in https://github.com/tektoncd/pipeline/issues/2134