[task-loops] Support string input for iterate parameter
When the input parameter is dynamically generated, it is not
possible to produce a list. Add support for newline separated
list of items passed a single string.

This makes it possible for a task to produce a result that
is than used as iterate parameter in the task loop, and thus
to programmatically define the list of tasks to be executed.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>