GitHub Notifier: Add support for Status API.
Using the Check Run API requires a GitHub App, and access to App debug details
(i.e. webhook delivery statuses, etc) require admin priviledges on the owning
user/org. For Tekton dogfooding, this makes using GitHub Apps difficult because:

1. For an org-internal App, we would need to grant anyone that would
need to debug dogfooding (i.e. build captains) admin access to the
tektoncd org, or put the burden on tektoncd admins to investigate
if/when there are issues with webhook delivery.
2. For an app owned by a separate org, we would need to make the App
public, which could let any GitHub user install the App and
cause us to receive their webhooks.

As an alternative, this PR adds support to make the behavior configurable based
on the authentication creds provided to the controller -
If a GitHub App ID + Key are passed in the Check Run API will be used, else we
fall back to the Status API and assume an access token has been provided to the
controller.