Send cloud events only, on every reconcile
Do not use the events package from pipeline as it will send both
k8s and cloud events. Send cloud events only instead.

The pipeline controller as enough context to decide whether to send
an event or not, based on previous and new condition. The cloud
events controller instead does not update the pipeline run condition,
so it is only aware of the current resource condition. The only
option for now is to send and event every time a resource is queued.

Pull in pipelines v0.24.3 to get https://github.com/tektoncd/pipeline/pull/3761
which fixes events on k8s 1.20.

Extend the cluster role to be able to write k8s events, so that we
may write an event in case of reconcile or cloud event delivery
failure.

Cloud events are now sent when a sink is specified. Note that since
resources do not have a condition when they're just started, the start
cloud event is not send for now. The controller will have to replicate
the logic in the pipelinerun controller to set a condition before the
event is sent.

Unit tests are passing, but coverage needs to be added.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>