[cloudevents] Use the cde sdk to create Events
Use the create methods provided by the SDK to create the events.
Re-use the objectWithCondition interface were possible to simplify
the logic of handling both TaskRuns and PipelineRuns (and Runs in
future).

The SDK does not set subject or link, so setting them additionally.
Set the status for pipelinerun events as per spec.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>