[cloudevents] Add support for handling artifact event types
Add a new module that implements artifact event types, with the
corresponding selection logic. Add to the send routine the ability
to loop among all available event creators, and send all matching
events.

Added test coverage for all the new modules and code.
Updated the SDK dependency to the latest available at the time
of commit.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>