[cloudevents] Allow for multiple annotations
Annotations is a dictionary so each key is unique.
Because of this, pipeline and tasks were previously limited to
one extra kind of cd event.

Change the notation to move the event type to the key, and thus
allow multiple events to be attached. The value is irrelevan for now
but it might be used in future to express extra configurations.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>