Provide git-repo and git-revision via annotations
There ist no other way to provide a commit-status-tracker target than
provide a PipelineResource of type git.
When I use the git-clone Task I cannot track the status of the commit.
So I implemented a way to provide the target git repository url and
target revision via annotations for the PipelineRun.

[commit-status-tracker]