[cloudevents] Fix source and subject
Remove the resource name from the source and make sure it's set
always as subject. The resource name in the source makes it harder
to setup event filters / triggers based on source, so it should not
be included in there. That information is available in the subject
anyways already.

The group/kind/version information is missing in the objects passed
to "ReconcileKind", so for now do not use them to build the source.
The kind (PipelineRun vs. TaskRun) can be identified from the object
type, group is always tekton.dev for Tekton resources, and the version
store is also always v1beta1, so they don't add any interesting info
to the source anyways.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>