Add instructions for new projects 📝
This isn't an exhaustive list but wanted to at least get it started!
I completely missed the presubmit-tests.sh update and @jerop caught it
for me but wanted to document this for folks in the future.

To be thorough we should also add more about how to add all the files
you need for your project (e.g. even that presubmit script assumes
certain other scripts are in the project folder)