Fix links to apache licenses 🔗
Not sure why these were trying to link to some kubernetes copy of the
license but that seems to be gone now, so instead just linking to the
existing license in the folder.

The markdown linter is a bit to vigorous and instead of just checking
the files I was changing decided to lint all the markdown files. Oh
well, that's one way to get it fixed I guess :D