Have the wait-task leverage `NewRequeueAfter`
This is a relatively new wrapped error in `knative.dev/pkg`, which automagically does the `EnqueueAfter`.

Had to rework some tests that were mocking `EnqueueAfter`, but the semantics should remain intact.

I also dropped the event this emits on successful reconciliation, since that can be a source of problems at scale.  Selfishly, this also reduced the amount of tweaking I had to do in the tests, so I figure win-win.