[task-loops] Repair Task Loops Nightly Builds
The Task Loops nightlies have been failing for a while (most recent example
here:
https://dashboard.dogfooding.tekton.dev/#/namespaces/tekton-nightly/pipelineruns/task-loops-release-nightly-s4qzd?pipelineTask=unit-tests&step=unit-test)

The source of the errors appear to be an extra `-mod=vendor` param in the
unit-tests task and the build task executing from the root of the repo instead
of the task-loops subdirectory.

This commit fixes both of these issues and should hopefully repair the
task-loops nightlies.