[remote-resolution] Proof-of-Concept Controller
This commit adds a controller, several reconcilers, and a set of helpers that
all work in tandem to provide a framework for resolving Tekton resources from
places like git repos. For this proof of concept only Tekton's Pipeline type is
resolvable but in principle the solutions included here are agnostic to the
content being requested.

In more concrete terms this project will take a PipelineRun that looks like
this:

```yaml
kind: PipelineRun
metadata:
  annotations:
    resolution.tekton.dev/resolver: git
    git.repo: https://github.com/sbwsg/experimental.git
    git.commit: 326ad1a1432b2144fb50f2bfd56875181a407c6b
    git.path: /remote-resolution/pipeline.yaml
spec:
  status: PipelineRunPending
```

and turn it in to one that looks like this:

```yaml
kind: PipelineRun
metadata:
  # <same as above>
status:
  # this pipelineSpec has been pulled in from git
  pipelineSpec:
    tasks:
    - name: fetch-from-git
      taskSpec:
        metadata: {}
        spec: null
        steps:
        - image: alpine/git
          name: ""
          resources: {}
          script: |
            git clone https://github.com/tektoncd/results
```

Here's a high level overview of the different parts in play:

1. (Shown in the YAML above) PipelineRuns get a new syntax to indicate the
place that a `pipelineSpec` should be fetched from. In this proof of concept
that syntax takes the form of some specific labels and annotations.

2. A component interprets the syntax from (1) and converts it into a generic
request for a resource. In this proof of concept the "generic request" is
either an HTTP call or an instance of a new CRD called a `ResourceRequest`.
Both have been implemented to try and figure out which approach better suits
the project's needs.

3. A resolver sees the request from (2) and figures out if it can respond. The
resolver fetches the resource if so. In this proof of concept three resolvers
have been implemented: a "noop" resolver for testing, an "in-cluster" resolver
that can fetch pipelines from inside the cluster, and a "git" resolver that can
fetch from public git repos.

The idea for (3) is to make writing a Resolver easy and to do so in a way that's
independent of the request format in (2). A resolver shouldn't care if the request
is coming in via HTTP or a CRD or anything else. To that end a go interface has
been introduced in the `pkg/reconciler/framework` directory that resolvers can
implement to hide all the boilerplate.

See remote-resolution/README.md for more details on installing and trying out
this controller.