[pipeline-to-taskrun] Support optional workspace ⚙️
Initially I left this out just to reduce the number of cases I needed to
cover but it was actually pretty easy to add. The functionality will be:
- When an optional workspace is provided, we do the same remapping we do
  for other workspace
- When an optional workspace is not provided, we don't do any remapping
  and just declare the same workspace as optional in the merged Task

I ran into this b/c I was trying to use pipeline to taskrun with the
latest git clone task in the catalog (0.4) which declares an optional
ssh-directory and I was frustrated that it didn't work.