Use cmpopts.IgnoreFields instead of IgnoreType.
From
https://pkg.go.dev/github.com/google/go-cmp/cmp/cmpopts#IgnoreTypes:

IgnoreTypes returns an Option that ignores all values assignable to certain types,
which are specified by passing in a value of each type.

This means when we pass in apis.Condition{}.LastTransitionTime.Inner.Time,
we are actually ignoring all time.Time, not just that particular field.
Similarly for apis.Condition{}.Message, we are ignoring all strings.

This change uses IgnoreFields to correct the behavior. Luckily, this doesn't
impact the tests in a meaningful way.