[Proof of Concept] Pipeline in a Pod
⚠ Poke around at your own risk; this code is not pretty! ⚠
I don't intend for this to be reviewed/merged, just to demonstrate what this feature would look like
and the level of effort required.

This commit adds a controller that reconciles PipelineRuns by implementing them in a pod.
(It doesn't work in a cluster with an existing Tekton Pipelines installation.)
Currently in Tekton, each task step is run in a container named "step-<step_name>".
This controller orders pipeline tasks sequentially and runs them with the existing entrypoint,
with each step run in a container named "task-<task_name>-step-<step_name>".
Parallel tasks, finally Tasks, runs, and task references aren't supported in this commit.
Features such as parameters, workspaces, results, sidecars, when expressions, and timeouts are also not supported.

The results of each step are written to the PipelineRun.Status.TaskRuns field.
If we decide to implement this feature in Pipelines, a better strategy could be to create TaskRuns and
update their statuses, and have the existing TaskRun controller ignore them via a label or similar.
Each reconcile loop would need to query for all TaskRuns managed by the PipelineRun, though not all of them
would need to be updated.