[github-app] Add support for repo-url Annotation in the TaskRun
Being able to add the repo url or just the url path
provides more flexibility and simplicity. One example
of this is when working with TriggerTemplates. The repo url
is simpler and more readily available to extract
from the payload than owner and repo.

This also moves the repo annotations to functions which
allows for better testing and validation.

Signed-off-by: Joe Stuart <jstuart@redhat.com>