Only initialise conditions for legacy events
When a resource first added to etcd, conditions are not set yet.
For feature parity with the tekton pipeline cloudevents, we shall
send a "started" event as soon as a pipeline run is stored in etcd
- as long as it's not "pending", so in that case we initialise
conditions for the pipeline run.

In all other cases we leave the resource as it is.
This means that for pipeline run we can send the CDEvents
pipelinerun.queued with no condition set, and the pipelinerun.started
once the pipeline has a condition set and thus started running.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>