[pipeline/trusted-resources]Verify task
This commit will introduce the validation of task, after verifying the
taskrun, the refered task will also be fetched and verified.
The second change is that webhook will make a copy of object and exclude
system populated fields. And verify on the copied object only on create
operation.

Before this commit, the webhook can only verify on taskspec.