[pipeline/trusted-resources]Verify task from cluster at taskrun
This commit focus on verify the task which is applied in cluster and
referred in taskrun. In this commit we will exclude system populated
fields and keep fields from users. The reason we verify the task in
taskrun validation not the task's validation is that there is a stage
between task creation and taskrun creation, so we need to make sure the
task is not tampered.

Before this commit we can only verify task from oci
bundle.