Create common release resources for custom tasks
This commit moves the release pipeline and task that are copy-pasted in many places in experimental/
to a common location. Individual projects in experimental should be migrated to use this Pipeline and Task
with appropriate parameter values, instead of continuing to copy paste these definitions in multiple places.
This will be used first for the pipeline-in-pod project (the ColocatedPipelineRun custom task).