[workflows] reconciler logic + docs
This commit implements the workflows reconciler. It creates Triggers for each Workflow
in the namespace of the Workflow, and gives the workflows EventListener access to
Triggers in all namespaces. Alternatively, we could create an EventListener per Workflow,
but that would require giving the workflows controller permission to create RBAC resources
in all namespaces.

Each Trigger created by a Workflow has an owner reference to the Workflow and a label
specifying the Workflow. The label is used by the Workflows reconciler to detect Triggers
belonging to the Workflow. Alternatively, we could add the managed Triggers to the Workflow
status.