[workflows] Implement filters
This commit adds filters to a workflow definition, which are converted
into ClusterInterceptors on a workflow's Triggers.
This commit adds support for a GitRef filter, which filters events to
those matching a git branch or tag matching a regex, and a custom
filter, which allows a user to specify CEL directly.
A PullRequest filter will be added in a later commit, after support
for GitHub events is implemented.