[workflows] Use PipelineRef from Tekton Pipelines
This commit updates the workflows spec to use a PipelineRef from
Tekton Pipelines, instead of introducing its own PipelineRefGit type.
One of the original goals of the workflows project was to allow storing
Tekton resources in version control, and this has now been implemented
in Pipelines. There's still some syntax simplification that could make
sense here, but for now we can start with a PipelineRef and iterate from there.