[PiP] Bump to pipeline v0.49.0.
The Pipelines-in-Pipelines controller is currently still depending
github.com/tektoncd/pipeline v0.20.0 and only works with the old Runs
resources. This means it basically can't be used with recent versions
of Tekton, as they use the new CustomRun resource.

This PR addresses this, by bumping this dependency to v0.49.0 and
change the use of v1alpha1.Run to v1beta1.CustomRun.